requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
pillow>=10.0.0
numpy>=1.24.0
pyyaml>=6.0

# Environment variables
python-dotenv>=1.0.0

# Google Gemini API (for teacher model)
google-generativeai>=0.3.0

# Text detection and OCR
python-doctr[torch]>=0.7.0  # DB-ResNet text detection
easyocr>=1.7.0
opencv-python>=4.8.0

# Metrics and evaluation
python-Levenshtein>=0.21.0

# Training utilities
tqdm>=4.65.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
peft>=0.7.0  # LoRA and parameter-efficient fine-tuning
sentencepiece>=0.1.99  # Tokenizer for Gemma
wandb>=0.16.0  # Experiment tracking (optional)

# Optional: For better performance
accelerate>=0.24.0
bitsandbytes>=0.41.0  # 8-bit quantization

# Development dependencies (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Alternative API support (optional)
# openai>=1.0.0  # For GPT models
# anthropic>=0.7.0  # For Claude models

datasets